requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-MySQLdb==0.2.0
gunicorn==18.0
idna==2.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.10
PyJWT==1.5.0
PyMySQL==0.7.1
PySocks==1.6.7
python-http-client==3.0.0
pytz==2017.2
requests==2.18.2
sendgrid==4.2.0
sendgrid-python==0.1.1
six==1.10.0
twilio==6.5.0
urllib3==1.22
Werkzeug==0.9.4